+++ app/src/main/java/com/example/eyespeak/SettingsButton.kt
@@ -2,38 +2,18 @@ package com.example.eyespeak
 
 import android.content.Context
 import android.content.Intent
-import android.speech.tts.TextToSpeech
 import android.util.AttributeSet
-import android.widget.Toast
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
 
+/**
+ * A subclass of VocableButton that will open SettingsActivity when interacted with
+ */
 class SettingsButton @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyle: Int = 0
-) : EyeSpeakButton(context, attrs, defStyle) {
+) : VocableButton(context, attrs, defStyle) {
 
-    init {
-        setOnClickListener {
-            navigateToSettings()
-        }
-    }
-
-    override fun onPointerEnter() {
-        buttonJob = backgroundScope.launch {
-            delay(2000)
-            uiScope.launch {
-                isPressed = true
-                Toast.makeText(context, text, Toast.LENGTH_LONG).show()
-                VocableTextToSpeech.getTextToSpeech()
-                    ?.speak(text, TextToSpeech.QUEUE_FLUSH, null, id.toString())
-                navigateToSettings()
-            }
-        }
-    }
-
-    private fun navigateToSettings() {
+    override fun performAction() {
         context.startActivity(Intent(context, SettingsActivity::class.java))
     }
 }
